Consolidating test cases as per PR suggestion
